refactor(test): audit regression.rs — dissolve file, rehome tests
- Delete test_out_of_range_value_at_start_tail_accepted (duplicate of
  pipeline.rs test_out_of_range_value_change_adopted)
- Move test_start_key_inclusion_value_checked → attack.rs as
  test_crafted_tampered_start_key_value_detected; simplify tampered_ops
- Move test_boundary_child_gap_closed_for_start_key → edge_cases.rs;
  fix misleading comment
- Rename db_a/db_b → source/target in moved tests
- Delete regression.rs and remove mod regression from mod.rs
