+++ firewood/src/merkle/tests/change/mod.rs
@@ -84,6 +84,5 @@ mod empty_and_bounded;
 mod fuzz;
 mod partial;
 mod pipeline;
-mod regression;
 mod root_hash;
 mod structural;
